Move some functions to common ChargePoint class
