+++ tests/conftest.py
@@ -24,7 +24,7 @@ def skip_notifications_fixture():
     with (
         patch("homeassistant.components.persistent_notification.async_create"),
         patch("homeassistant.components.persistent_notification.async_dismiss"),
-        patch("custom_components.ocpp.api.ChargePoint.notify_ha"),
+        patch("custom_components.ocpp.chargepoint.ChargePoint.notify_ha"),
     ):
         yield
